Rename package to optimade_client
Reset version to 1.0.0.
