+++ README.md
@@ -1,7 +1,7 @@
-# OPTIMADE client (in Jupyter)
+# OPTIMADE client (in Voilà)
 
-[![Materials Cloud](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/aiidalab/aiidalab-optimade/v3/docs/resources/mcloud_badge.json)](https://dev-tools.materialscloud.org/optimadeclient/)
-[![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/aiidalab/aiidalab-optimade/v3?urlpath=%2Fvoila%2Frender%2FOPTIMADE%20Client.ipynb)
+[![Materials Cloud](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/CasperWA/voila-optimade-client/v3/docs/resources/mcloud_badge.json)](https://dev-tools.materialscloud.org/optimadeclient/)
+[![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/CasperWA/voila-optimade-client/v3?urlpath=%2Fvoila%2Frender%2FOPTIMADE%20Client.ipynb)
 
 Query for and import structures from [OPTIMADE](https://www.optimade.org) providers (COD, Materials Cloud, NoMaD, Materials Project, ODBX, OQMD, and more ...).
 
@@ -13,7 +13,7 @@ This Jupyter-based app is intended to run either:
 
 - In [AiiDA lab](https://aiidalab.materialscloud.org) as well as inside a [Quantum Mobile](https://materialscloud.org/work/quantum-mobile) Virtual Machine;
 - As a [Materials Cloud tool](https://dev-tools.materialscloud.org/optimadeclient/);
-- As a standalone [MyBinder application](https://mybinder.org/v2/gh/aiidalab/aiidalab-optimade/v3?urlpath=%2Fvoila%2Frender%2FOPTIMADE%20Client.ipynb); or
+- As a standalone [MyBinder application](https://mybinder.org/v2/gh/CasperWA/voila-optimade-client/v3?urlpath=%2Fvoila%2Frender%2FOPTIMADE%20Client.ipynb); or
 - As a standalone local application (see more information about this below).
 
 For AiiDA lab, Use the App Store in the [Home App](https://github.com/aiidalab/aiidalab-home) to install it.
@@ -25,7 +25,7 @@ For AiiDA lab, Use the App Store in the [Home App](https://github.com/aiidalab/a
 To use the OPTIMADE structure importer in your own AiiDA lab application write the following:
 
 ```python
-from aiidalab_optimade import OptimadeQueryWidget
+from optimade_client import OptimadeQueryWidget
 from aiidalab_widgets_base.viewers import StructureDataViewer
 from ipywidgets import dlink
