+++ optimade_client/informational.py
@@ -6,13 +6,13 @@ from urllib.parse import urlencode
 
 import ipywidgets as ipw
 
-from aiidalab_optimade.logger import LOGGER, WIDGET_HANDLER, REPORT_HANDLER
+from optimade_client.logger import LOGGER, WIDGET_HANDLER, REPORT_HANDLER
 
-from aiidalab_optimade.utils import __optimade_version__
+from optimade_client.utils import __optimade_version__
 
 
 IMG_DIR = Path(__file__).parent.parent.joinpath("img")
-SOURCE_URL = "https://github.com/aiidalab/aiidalab-optimade/"
+SOURCE_URL = "https://github.com/CasperWA/voila-optimade-client/"
 
 
 class HeaderDescription(ipw.VBox):
