+++ optimade_client/subwidgets/results.py
@@ -4,8 +4,8 @@ from urllib.parse import urlparse, parse_qs
 import ipywidgets as ipw
 import traitlets
 
-from aiidalab_optimade.exceptions import InputError
-from aiidalab_optimade.logger import LOGGER
+from optimade_client.exceptions import InputError
+from optimade_client.logger import LOGGER
 
 
 __all__ = ("StructureDropdown", "ResultsPageChooser")
