+++ optimade_client/utils.py
@@ -17,11 +17,11 @@ import requests
 from optimade.models import LinksResource, OptimadeError, Link, LinksResourceAttributes
 from optimade.models.links import LinkType
 
-from aiidalab_optimade.exceptions import (
+from optimade_client.exceptions import (
     ApiVersionError,
     InputError,
 )
-from aiidalab_optimade.logger import LOGGER
+from optimade_client.logger import LOGGER
 
 
 # Supported OPTIMADE spec versions
